refactor: changed folder structure for utilities
