+++ src/utilities/FilterUtilities/getCategoryFilteredProducts.js
