+++ src/utilities/index.js
@@ -1,5 +1,5 @@
-export { getSortedProducts } from "./getSortedProducts";
-export { getCategoryFilteredProducts } from "./getCategoryFilteredProducts";
-export { getPriceFilteredProducts } from "./getPriceFilteredProducts";
-export { getRatingFilteredProducts } from "./getRatingFilteredProducts";
-export { getFilteredProducts } from "./getFilteredProducts";
+export { getSortedProducts } from "./FilterUtilities/getSortedProducts";
+export { getCategoryFilteredProducts } from "./FilterUtilities/getCategoryFilteredProducts";
+export { getPriceFilteredProducts } from "./FilterUtilities/getPriceFilteredProducts";
+export { getRatingFilteredProducts } from "./FilterUtilities/getRatingFilteredProducts";
+export { getFilteredProducts } from "./FilterUtilities/getFilteredProducts";
